Avoid importing 'util' (no UTF-8 Unicode encoding on Node <=11)+lint eslint conf
This is due to this paradox that I'm still not sure is solvable.

If you use a bundler to include this file, which is supposed to build
for the browser, it'd take the CJS path which assumes Node doesn't have
TextEn/Decoder and imports it from util. Problem is that the bundler
does too, and either includes an enormous polyfill at best, or complains
it doesn't exist at worst.

I couldn't find any way to detect Node or not because the bundlers do
that inclusion statically, despite the amount of checking I did for
process.version etc. Soooo... had to come up with this compromise to
simply NOT! support encoding Unicode UTF-8 in Node versions older than
11. Oh no, right?

Also:
* make syntax changes in test.js that make eslint run on it, so I can
actually lint it in the next commit.
* lint eslint.config.cjs so that it doesn't look as jarring when I
update it next commit